Optimize map loading (#56)
* optimize map loading

* Make rvalue constraint clearer

* Make decode safer

* Add base64 tests

* Update Makefile

+++ src/core/spatial/region_tree/node_traits.h
@@ -9,10 +9,10 @@ class NodeTraits {
     static constexpr std::size_t MAX_CHILDREN = 25;
 };
 
-// RegionTree<size_t> is used for testing purposes
+// RegionTree<int> is used for testing purposes
 // MAX_CHILDREN is set to 4 to make testing easier
 template<>
-class NodeTraits<size_t> {
+class NodeTraits<int> {
   public:
     static constexpr std::size_t MAX_CHILDREN = 4;
 };
